requirements.txt
feedparser>=6.0
requests>=2.31
google-genai>=1.0
pyyaml>=6.0
jinja2>=3.1
python-dotenv>=1.0
beautifulsoup4>=4.12
